Added All Assignmnts
